requirements.txt
Django>=5.0.0,<6.0.0
python-dotenv>=1.0.0
Jinja2>=3.1.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
PyJWT>=2.8.0
pymongo>=4.6.0
sqlparse>=0.4.4
groq>=0.4.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0
certifi>=2024.2.2
reportlab>=4.0.0
Pillow>=10.0.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
